Add files via upload
The response of the treatment was measured as reduction of tumour size. It is pending to assess if that treatment also reduced the number of metastatic sites. This could be done by creating the same box plot but comparing mean number of metastatic sites across different treatments. This would be interesting as possibly the treatment reduces the tumor size but has no effect in the number of metastatic sites.